Damage modifiers 2

+++ LuckParser/Models/JsonModels/JsonBuffDamageModifierData.cs
@@ -24,7 +24,7 @@ namespace LuckParser.Models.JsonModels
             /// <summary>
             /// Gained damage
             /// </summary>
-            public int DamageGain;
+            public double DamageGain;
             /// <summary>
             /// Total damage done
             /// </summary>
